refactor: GameService 리팩토링 - Custom Exception, State Pattern onExit, GameQueryService 분리
- RuntimeException 6개를 ErrorCode 기반 Custom Exception으로 교체
- GamePhaseState에 onExit() 추가, 5개 State 클래스에 구현
- PhaseResultProcessor 신규 생성 (투표/밤행동/Redis 동기화 로직 추출)
- processCurrentPhase() switch문 제거 → State Pattern 위임
- GameQueryService 신규 생성 (조회 메서드 10개 추출)
- ChatRoomService → GameQueryService 의존성 변경 (순환 참조 해소)
- GameController 조회 호출 GameQueryService로 변경
- Deprecated SchedulerTimerService.startTimer(String) 제거
- 미사용 import/field 정리
