+++ src/main/java/com/example/mafiagame/game/state/DayFinalDefenseState.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Component;
 
 import com.example.mafiagame.game.domain.state.GamePhase;
 import com.example.mafiagame.game.domain.state.GameState;
+import com.example.mafiagame.game.service.PhaseResultProcessor;
 
 /**
  * 최후 변론 페이즈 상태
@@ -16,6 +17,11 @@ public class DayFinalDefenseState implements GamePhaseState {
     @Override
     public void process(GameState gameState) {
         gameState.setGamePhase(GamePhase.DAY_FINAL_DEFENSE);
+    }
+
+    @Override
+    public void onExit(GameState gameState, PhaseResultProcessor processor) {
+        // 최후 변론 종료 시 별도 처리 없음
     }
 
     @Override
