+++ src/main/java/com/example/mafiagame/global/error/ErrorCode.java
@@ -12,6 +12,9 @@ public enum ErrorCode {
     GAMESTATE_NOT_FOUND(HttpStatus.NOT_FOUND, "GAMESTATE_NOT_FOUND", "Game State Not Found"),
     USER_NOT_FOUND(HttpStatus.NOT_FOUND, "USER_NOT_FOUND", "User Not Found"),
     USER_ALREADY_EXISTS(HttpStatus.BAD_REQUEST, "USER_ALREADY_EXISTS", "User Already Exists"),
+    INSUFFICIENT_PLAYERS(HttpStatus.BAD_REQUEST, "INSUFFICIENT_PLAYERS", "게임을 시작하려면 최소 4명이 필요합니다."),
+    GAME_CREATE_IN_PROGRESS(HttpStatus.CONFLICT, "GAME_CREATE_IN_PROGRESS", "게임 생성 중입니다. 잠시 후 다시 시도해주세요."),
+    GAME_CREATE_INTERRUPTED(HttpStatus.INTERNAL_SERVER_ERROR, "GAME_CREATE_INTERRUPTED", "게임 생성 중 인터럽트가 발생했습니다."),
     ;
 
     private final String code;
